build: upgrade to Angular v11 and switch to ng-packagr
To avoid the redundancy of a secondary build system that only
degrades over time, we switch to ng-packagr, since it stays true
to the latest revision of APF (Angular Package Format).

Bump latest dependencies to Angular v11 (and necessary peers),
add Ivy processing for test and e2e, and add necessary typing
to PrebootModule for generic compatibility.