refactor: remove tsickle and configure library generation
Tsickle is intended to be used for libraries intended to be compiled with Closure Compiler.

This is not recommanded to be used outside of Google as also the Angular framework is no longer closure compatable by default via NPM.

Also, with this change we change the library to be compiled with VE since, we shouldn't ship libraries in Ivy mode.